+++ lib/rust/ensogl/core/src/system/gpu/data/attribute.rs
@@ -218,7 +218,8 @@ impl {
     pub fn update(&mut self) {
         debug_span!("Updating.").in_scope(|| {
             if self.used_size * 2 < self.size() {
-                self.shrink_to_fit();
+                // FIXME (#6340)
+                //self.shrink_to_fit();
             }
             if self.shape_dirty.check() {
                 for i in 0..self.buffers.len() {